remove preview that con no longer be doen without public serialization support. fix some things that broke through merges
remove serialization dependency
