+++ build.gradle.kts
@@ -27,8 +27,6 @@ plugins {
     @Suppress("DSL_SCOPE_VIOLATION")
     alias(libs.plugins.gradle.secrets) apply false
     @Suppress("DSL_SCOPE_VIOLATION")
-    alias(libs.plugins.kotlin.serialization) apply false
-    @Suppress("DSL_SCOPE_VIOLATION")
     alias(libs.plugins.kapt) apply false
     @Suppress("DSL_SCOPE_VIOLATION")
     alias(libs.plugins.ksp) apply false
