+++ microapps/FeatureFormsApp/app/src/main/java/com/arcgismaps/toolkit/featureformsapp/MainActivity.kt
@@ -53,4 +53,3 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
