+++ toolkit/featureforms/build.gradle.kts
@@ -2,7 +2,6 @@ plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
     id("artifact-deploy")
-    id("kotlinx-serialization")
 }
 
 android {
